app/requirements.txt
alembic==1.7.6
aniso8601==9.0.1
boto3==1.20.54
botocore==1.23.54
coverage==6.3.2
click==8.0.3
Flask==2.0.2
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
importlib-metadata==4.11.0
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
Mako==1.1.6
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.27.0
numpy==1.22.2
Pillow==9.0.1
psycopg2-binary==2.9.3
pytest==7.1.2
pytest-mock==3.7.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
s3transfer==0.5.1
six==1.16.0
SQLAlchemy==1.4.31
urllib3==1.26.8
Werkzeug==2.0.3
zipp==3.7.0
